dockerized django

+++ my_account/models.py
@@ -80,7 +80,7 @@ class User(AbstractBaseUser,PermissionsMixin):
     middle_name = models.CharField(_("middle name"),max_length=40,blank=True,null=True)
     last_name = models.CharField(_("last name"),max_length=40)
     email = models.EmailField(_("email"),max_length=200,unique=True,blank=True,null=True)
-    phone_number = models.IntegerField(_("phone number"),unique=True)
+    phone_number = models.BigIntegerField(_("phone number"),unique=True)
     is_email_verified = models.BooleanField(_("email verified"),default=False)
     address = models.CharField(_("address"),max_length=100,blank=True,null=True)
     has_verified_dairy = models.BooleanField(_("has verified dairy"),default=False)
